*extracted class Parser
